feat(stacks): Add stack interface and NodeJS implementation
The stack is an interface that is used to implement how each stack of a project is detected (and
modified in a future implementation). The first method is "existsInPath" which will search in a
given path for files or specific aspects of the stack and tell us whether the stack is present in
the project.