docs(rules-doc): Update doc for node JS rule
Update the doc because this class will not have the same purpose as
before. It is now a rule that only applies to Semver notations in
package.json

BREAKING CHANGE: This class will not cover every aspect of node js anymore. It is now a rule for semver in package.json