fix(ExactNpmVersion Rule): Register for specific rules, remove check of file
There was a check for package.json existence, but it is no longer
necessary as the stacks which this rule is registered for, already check
for this file. For example, if the Node and TypeScript stacks are not
found because of the absence of package.json, this rule will not be
called.