feat(Release): Update to version 0.7.0
We now use a substack hierarchy for our stacks. For example, it is
possible to have substacks of a stack, like Node Backend is a substack
of Node. A new rule for nodemon has been added. The rootPath variable is
not passed to each class anymore, instead we use a static variable in
Globals class. The Node Stack has been refactored.